+++ packages/passport/src/constants/error.ts
@@ -22,6 +22,8 @@ export const NOT_FOUND_VALID_CLIENT_ID = 'Not found valid clientId. clientId opt
 
 export const NOT_FOUND_ID_TOKEN = 'Not found valid id token';
 
+export const NOT_FOUND_REFRESH_TOKEN_ENTRY = 'Not found cached refresh token entry';
+
 export const REFRESH_TOKEN_EXPIRED = 'Current refresh token in memory expired. You need to perform re-authentication';
 
 export const AUTHENTICATION_ACCESS_DENIED =
@@ -36,8 +38,12 @@ export const INVALID_DECODED_ID_TOKEN = 'Current id token could not be decoded';
 
 export const INVALID_BASE64_STRING = 'Base64 string is not of the correct length';
 
-export const INVALID_WEB_WORKER_INSTANCE = 'Auth worker must be instance of ""WebWorker""';
+export const INVALID_TOKEN_ROTATION = 'Token rotation is not valid status';
 
 export const REMAIN_PREV_TRANSACTION_INFO = 'Prev transaction info is remained in storage';
 
 export const NOT_AUTHENTICATED_USER = 'Current user is not authenticated. Please do re-authentication process';
+
+export const AUTHORIZATION_CODE_FLOW = 'Currently in code transaction phase, therefore access is not available';
+
+export const INVALID_CACHE_LOCATION = 'Invalid cache location';