+++ packages/passport/src/constants/index.ts
@@ -8,7 +8,13 @@ export const CACHE_LOCATION_COOKIE = 'cookie';
 
 export const TRANSACTION_STORAGE_KEY_PREFIX = 'passport.transaction.data';
 
-export const CACHE_STORAGE_KEY_PREFIX = 'passport.cache.data';
+export const CACHE_KEY_PREFIX = 'passport.cache.data';
+
+export const CACHE_KEY_AUTHENTICATED_SUFFIX = 'authenticated';
+
+export const CACHE_KEY_ID_TOKEN_SUFFIX = 'idToken';
+
+export const CACHE_KEY_REFRESH_TOKEN_SUFFIX = 'refreshToken';
 
 export const DEFAULT_PASSPORT_PREFIX_POS = 0;
 
@@ -17,3 +23,5 @@ export const AUTHORIZATION_RESPONSE_TYPE = 'code';
 export const CODE_CHALLENGE_METHOD = 'S256';
 
 export const DEFAULT_MIN_VALIDITY = 1;
+
+export const CACHE_LOCATION_MEMORY = 'memory';